docs/implementation_plan/phase_3_power_spectrum/step_3.2_subpeaks_analysis/subpeaks_analyzer.py
"""
High-l sub-peaks analysis module.

Analyzes sub-peaks from ω_min beatings and detects predicted
l≈4500-6000 peak.

Author: Vasiliy Zdanovskiy
Email: vasilyvz@gmail.com
"""

from typing import Optional, List, Dict, Tuple
import numpy as np
from dataclasses import dataclass
from scipy.signal import find_peaks
from cmb.spectrum.power_spectrum import PowerSpectrum
from cmb.theta_evolution_processor import ThetaEvolutionProcessor
from config.settings import Config


@dataclass
class SubPeak:
    """
    Sub-peak information.

    Attributes:
        l_position: Multipole position of peak
        amplitude: Peak amplitude
        width: Peak width
        significance: Statistical significance
        beating_match: Corresponding beating from ω_min(t)
    """

    l_position: float
    amplitude: float
    width: float
    significance: float
    beating_match: Optional[float]


@dataclass
class SubPeakAnalysis:
    """
    Sub-peak analysis results.

    Attributes:
        peaks: List of identified sub-peaks
        high_l_peak: High-l peak (l≈4500-6000) if found
        beating_frequencies: Calculated beating frequencies
        analysis_metadata: Additional analysis metadata
    """

    peaks: List[SubPeak]
    high_l_peak: Optional[SubPeak]
    beating_frequencies: np.ndarray
    analysis_metadata: dict


class SubpeaksAnalyzer:
    """
    High-l sub-peaks analyzer.

    Identifies sub-peaks from ω_min beatings and detects
    predicted high-l peak.
    """

    def __init__(
        self,
        power_spectrum: PowerSpectrum,
        evolution_processor: ThetaEvolutionProcessor,
    ):
        """
        Initialize analyzer.

        Args:
            power_spectrum: Power spectrum data
            evolution_processor: Θ-field evolution processor
        """
        self.power_spectrum = power_spectrum
        self.evolution_processor = evolution_processor

    def analyze_subpeaks(self) -> SubPeakAnalysis:
        """
        Analyze sub-peaks in power spectrum.

        Identifies sub-peaks from beatings and detects high-l peak.

        Returns:
            SubPeakAnalysis instance with results

        Raises:
            ValueError: If analysis fails
        """
        pass

    def _calculate_beatings(self) -> np.ndarray:
        """
        Calculate beating frequencies from ω_min(t) evolution.

        Beatings arise from modulations in ω_min(t) temporal evolution.
        Each beating frequency corresponds to a sub-peak in power spectrum.

        Algorithm:
        1. Analyze ω_min(t) evolution for modulations
        2. Extract beating frequencies from modulations
        3. Convert to multipoles using l = π D ω

        Returns:
            Array of beating frequencies (in Hz)

        Raises:
            ValueError: If evolution data is insufficient
        """
        pass

    def _find_subpeaks(self) -> List[SubPeak]:
        """
        Find sub-peaks in power spectrum.

        Returns:
            List of identified sub-peaks
        """
        pass

    def _detect_high_l_peak(
        self, l_min: float = 4500.0, l_max: float = 6000.0
    ) -> Optional[SubPeak]:
        """
        Detect high-l peak in specified range.

        Args:
            l_min: Minimum multipole for search
            l_max: Maximum multipole for search

        Returns:
            SubPeak if found, None otherwise
        """
        pass

    def _match_beatings_to_peaks(
        self, peaks: List[SubPeak], beatings: np.ndarray
    ) -> List[SubPeak]:
        """
        Match beatings to sub-peak positions.

        Args:
            peaks: List of sub-peaks
            beatings: Beating frequencies

        Returns:
            List of sub-peaks with beating matches
        """
        pass
