requirements.txt
# Core dependencies for CMB verification project
# Author: Vasiliy Zdanovskiy
# Email: vasilyvz@gmail.com

# Numerical computing
numpy>=2.0.0

# HEALPix operations
healpy>=1.19.0

# Astronomical data handling
astropy>=7.0.0

# Visualization
matplotlib>=3.10.0

# Configuration and data formats
pyyaml>=6.0.0

# Code quality tools
black>=25.0.0
flake8>=7.0.0
mypy>=1.19.0
pytest>=9.0.0


